requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.17

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database ORM
tortoise-orm==0.21.6
aerich==0.7.2

# OpenAI API
openai>=1.50.0

# Environment variables
python-dotenv==1.0.1

# HTTP client
aiohttp>=3.9.0
httpx>=0.27.0

# Data validation
pydantic<2.10,>=2.4.1
pydantic-settings<2.7,>=2.0

# PDF processing
PyPDF2>=3.0.0

# Web scraping
beautifulsoup4>=4.12.0

